requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.7
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-search-results==2.4.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
mistralai==0.4.2
orjson==3.10.6
packaging==24.1
ply==3.11
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
replicate==0.30.1
repoze.lru==0.7
requests==2.32.3
rsa==4.9
simplegeneric==0.8.1
sniffio==1.3.1
sqlparse==0.5.1
statsd==4.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
